requirements.txt
# Web UI
streamlit==1.28.0

# LangChain and vectorstores
langchain==0.2.20
langchain-community==0.0.199
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# LLM / API client
openai==1.31.0
python-dotenv==1.0.1

# PDF loaders
PyMuPDF==1.22.5

# Optional / common utils
regex==2025.9.15
